+++ src/optics/Optics.jl
@@ -18,7 +18,9 @@ export AbstractOpticalProps,
     TwoStream,
     compute_col_gas!,
     compute_relative_humidity!,
-    compute_optical_props!
+    compute_optical_props!,
+    add_cloud_optics_lw!,
+    add_cloud_optics_sw!
 
 """
     AbstractOpticalProps